Move testhelpers into separate package
