+++ tests/new_integration_tests/test_issues.py
@@ -5,8 +5,8 @@ import tempfile
 import collections
 
 import repobee_plug as plug
-from repobee_plug.testhelpers import funcs
-from repobee_plug.testhelpers import const
+from repobee_testhelpers import funcs
+from repobee_testhelpers import const
 
 _TestIssue = collections.namedtuple("_TestIssue", "title body path")
